fix: reset game and show intro screen upon resizing window Fixes #11
